+++ src/Objects/PulseColumnTimelineValue.php
@@ -17,7 +17,7 @@ class PulseColumnTimelineValue extends PulseColumnValue
      *
      * @since  0.2.1
      *
-     * @return \DateTime[] The timeline's begin and end dates
+     * @return \DateTime[]|null The timeline's begin and end dates
      *
      *     array(
      *       'from' => \DateTime
@@ -26,12 +26,17 @@ class PulseColumnTimelineValue extends PulseColumnValue
      */
     public function getValue ()
     {
+        if ($this->isNullValue())
+        {
+            return null;
+        }
+
         if (!isset($this->column_value))
         {
-            $this->column_value = array(
+            $this->column_value = [
                 'from' => new \DateTime($this->jsonResponse["value"]["from"]),
                 'to'   => new \DateTime($this->jsonResponse["value"]["to"])
-            );
+            ];
         }
 
         return $this->column_value;
@@ -45,22 +50,30 @@ class PulseColumnTimelineValue extends PulseColumnValue
      * @param \DateTime $from
      * @param \DateTime $to
      *
+     * @since 0.3.0 \InvalidArgumentException is now thrown
      * @since 0.2.1
+     *
+     * @throws \InvalidArgumentException if $from or $to are not \DateTime instances
      */
     public function updateValue ($from, $to)
     {
+        if (!($from instanceof \DateTime) || !($to instanceof \DateTime))
+        {
+            throw new \InvalidArgumentException('$from and $to are expected to be \\DateTime instances');
+        }
+
         $url        = sprintf("%s/%d/columns/%s/timeline.json", self::apiEndpoint(), $this->board_id, $this->column_id);
-        $postParams = array(
+        $postParams = [
             "pulse_id" => $this->pulse_id,
             "from"     => $from->format('Y-m-d'),
             "to"       => $to->format('Y-m-d')
-        );
+        ];
 
         self::sendPut($url, $postParams);
 
-        $this->column_value = array(
+        $this->column_value = [
             'from' => $from,
             'to'   => $to
-        );
+        ];
     }
 }